SAColaoECommerce/requirements/base.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-environ==0.4.4
djangorestframework==3.8.2
idna==2.6
pkg-resources==0.0.0
pytz==2018.4
requests==2.18.4
Serializer==0.2.1
simplejson==3.15.0
six==1.11.0
urllib3==1.22
